Fixed making proper uppercase letters lowercase to fix sheep names
This wouldn't have been a problem because there is a uppercase check for spaces, but there isn't one for brackets. So letters after a opening bracket weren't automatically capitalized and thus made lowercase. In this case (sheep color names in brackets started with lowercase) the original color names were already properly capitalized, but were made lowercase. I couldn't think of a situation where the toLowerCase() did something beneficial (the letters after the capital one are usually already lowercase), but one could re-add it if there was such a situation.