Overhauled mob view (#246)
* Sorted mobs by their names instead of by their resourceLocations

* Fixed Blocks being lit up from above instead of the standard lighting

(for now)
As the entity render isn't working, the line I removed only had an effect on the blocks dropped from certain mobs. That lighting didn't feel natural at all and also caused some JEI tab icons to change lighting when no drop is present.

* Removed a Sheep from the mob view that dropped no wool at all

* Fixed colored sheep names in other languages being untranslated

I know, this is not an improvement for the english translation, but in other languages this solution is in my opinion still better than having an untranslated bit before a translated one in a mob name.

* Added the experience values for the vanilla mobs

* Added the LightLevels to some Vanilla mobs

* Fixed making proper uppercase letters lowercase to fix sheep names

This wouldn't have been a problem because there is a uppercase check for spaces, but there isn't one for brackets. So letters after a opening bracket weren't automatically capitalized and thus made lowercase. In this case (sheep color names in brackets started with lowercase) the original color names were already properly capitalized, but were made lowercase. I couldn't think of a situation where the toLowerCase() did something beneficial (the letters after the capital one are usually already lowercase), but one could re-add it if there was such a situation.

* Increased revis number

* Move xp and LightLevel definition to MobCompat

This also reverts the two "Added" commits.

* Made ResourceLocation the key in certain maps again and enhanced comment

This also reverts some changes that were made in the first commit of this PR.

* Made MobCompat a Singleton class

* Removed some imports and a testing remnant, added curly brackets and...

...moved long if-condition to its own function

* Replaced Integer[] with Tuple<Integer, Integer>